bugfix: security 권한 필요로 인한 테스트 코드 api 위치 변경
